+++ app/models/metadata_presenter/autocomplete_item.rb
@@ -1,0 +1,9 @@
+class MetadataPresenter::AutocompleteItem < MetadataPresenter::Metadata
+  def id
+    value
+  end
+
+  def name
+    text
+  end
+end