+++ config/initializers/supported_components.rb
@@ -21,7 +21,7 @@ Rails.application.config.supported_components =
       content: %w(content)
     },
     singlequestion: {
-      input: %w(text textarea number date radios checkboxes email upload),
+      input: %w(text textarea number date radios checkboxes email upload autocomplete),
       content: %w()
      }
   })